requirements.txt
# ============================================================
# StudyMate AI — Requirements
# Python 3.11.9  (pinned via .python-version file)
#
# KEY RULES:
#  1. --extra-index-url must be on its OWN line before torch
#  2. torch has NO version pin — pip picks correct version
#  3. sentence-transformers pinned to 2.7.0 (last pre-torch2 version)
#  4. faiss-cpu has no CUDA deps — keeps RAM low
# ============================================================

# PyTorch CPU-only  ← index URL on separate line, no +cpu suffix
--extra-index-url https://download.pytorch.org/whl/cpu
torch

# Web Framework
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
aiosqlite

# LLM — Groq via OpenAI-compatible SDK
openai

# Embeddings — smallest model, lazy-loaded, CPU-only
sentence-transformers==2.7.0

# Vector Store
faiss-cpu
numpy

# PDF
pdfplumber
PyMuPDF

# YouTube
youtube-transcript-api
yt-dlp

# Audio
gTTS

# Text
tiktoken

# Utilities
python-dotenv
pydantic
pydantic-settings
httpx
aiofiles
requests
tenacity
loguru
rich
tqdm
